+++ plotnine/scales/scale_manual.py
@@ -1,5 +1,7 @@
 from warnings import warn
 
+import numpy as np
+
 from ..doctools import document
 from ..exceptions import PlotnineWarning
 from ..utils import alias
@@ -16,6 +18,15 @@ class _scale_manual(scale_discrete):
     {superclass_parameters}
     """
     def __init__(self, values, **kwargs):
+        # Match the values of the scale with the breaks (if given)
+        if 'breaks' in kwargs:
+            breaks = kwargs['breaks']
+            if np.iterable(breaks) and not isinstance(breaks, str):
+                if iter(breaks) is breaks:
+                    breaks = list(breaks)
+                    kwargs['breaks'] = breaks
+                values = {b: v for b, v in zip(breaks, values)}
+
         self._values = values
         scale_discrete.__init__(self, **kwargs)
 
@@ -36,7 +47,8 @@ class scale_color_manual(_scale_manual):
     Parameters
     ----------
     values : array_like
-        Colors that make up the palette.
+        Colors that make up the palette. The values will be matched with
+        the ``limits`` of the scale or the ``breaks`` if provided.
     {superclass_parameters}
     """
     _aesthetics = ['color']
@@ -50,7 +62,8 @@ class scale_fill_manual(_scale_manual):
     Parameters
     ----------
     values : array_like
-        Colors that make up the palette.
+        Colors that make up the palette. The values will be matched with
+        the ``limits`` of the scale or the ``breaks`` if provided.
     {superclass_parameters}
     """
     _aesthetics = ['fill']
@@ -66,7 +79,8 @@ class scale_shape_manual(_scale_manual):
     values : array_like
         Shapes that make up the palette. See
         :mod:`matplotlib.markers.` for list of all possible
-        shapes.
+        shapes. The values will be matched with the ``limits``
+        of the scale or the ``breaks`` if provided.
     {superclass_parameters}
 
     See Also
@@ -100,6 +114,8 @@ class scale_linetype_manual(_scale_manual):
             2. Tuples of the form (offset, (on, off, on, off, ....))
                e.g. (0, (1, 1)), (1, (2, 2)), (2, (5, 3, 1, 3))
 
+        The values will be matched with the ``limits`` of the scale
+        or the ``breaks`` if provided.
     {superclass_parameters}
 
     See Also
@@ -126,7 +142,8 @@ class scale_alpha_manual(_scale_manual):
     ----------
     values : array_like
         Alpha values (in the [0, 1] range) that make up
-        the palette.
+        the palette. The values will be matched with the
+        ``limits`` of the scale or the ``breaks`` if provided.
     {superclass_parameters}
     """
     _aesthetics = ['alpha']
@@ -140,7 +157,8 @@ class scale_size_manual(_scale_manual):
     Parameters
     ----------
     values : array_like
-        Sizes that make up the palette.
+        Sizes that make up the palette. The values will be matched
+        with the ``limits`` of the scale or the ``breaks`` if provided.
     {superclass_parameters}
     """
     _aesthetics = ['size']